Enable 'treat warnings as errors' on all projects.
We appeared to have a string using `$(WarningsAsErrors)` in the 'list of warnings that are treated as errors' but it would still treat most warnings as warnings.

I just enabled treat all warnings as errors for each project which is fine as we don't have any warnings at the moment.